MdePkg: Add Storage Device Class in Boot Manager Policy
UEFI 2.11 Specification adds storage class devices in the
EFI_BOOT_MANAGER_POLICY_CONNECT_DEVICE_CLASS method of the
EFI_BOOT_MANAGER_POLICY protocol.
This is required for scenarios where the boot loader or BDS needs to
connect storage devices to search for specific partition or file
without having to take the hit of a connect all.
This further allows for platform policy to enumerate a limited number of
devices in more vertically integrated purpose-built devices.

Signed-off-by: Sachin Ganesh <sachinganesh@ami.com>